fix some errors, fix some layout issues (tables)
