+++ src/comp/elements/NostrReceiveQR.svelte
@@ -1,0 +1,59 @@
+<script lang="ts">
+	import { browser } from '$app/environment';
+	import { nip19 } from 'nostr-tools';
+	import { QRCodeImage } from 'svelte-qrcode-image';
+	import { nostrPubKey, useExternalNostrKey, useNostr } from '../../stores/nostr';
+	const getEncodedNpub = async () => {
+		if (!$useNostr) {
+			return '';
+		}
+		if ($useExternalNostrKey) {
+			if (browser) {
+				const pubK = await window.nostr?.getPublicKey();
+				if (!pubK) {
+					return '';
+				}
+				return nip19.npubEncode(pubK);
+			}
+			return '';
+		} else {
+			if (!$nostrPubKey) {
+				return '';
+			}
+			return Promise.resolve(nip19.npubEncode($nostrPubKey));
+		}
+	};
+</script>
+
+{#await getEncodedNpub() then npub}
+	{#if $useNostr && npub}
+		<!-- content here -->
+		<div class="flex items-center justify-center w-full">
+			<div class="flex flex-col gap-2 items-center">
+				<div class="badge badge-info gap-2">
+					<svg
+						xmlns="http://www.w3.org/2000/svg"
+						fill="none"
+						viewBox="0 0 24 24"
+						stroke-width="1.5"
+						stroke="currentColor"
+						class="w-3 h-3"
+					>
+						<path
+							stroke-linecap="round"
+							stroke-linejoin="round"
+							d="M6.827 6.175A2.31 2.31 0 015.186 7.23c-.38.054-.757.112-1.134.175C2.999 7.58 2.25 8.507 2.25 9.574V18a2.25 2.25 0 002.25 2.25h15A2.25 2.25 0 0021.75 18V9.574c0-1.067-.75-1.994-1.802-2.169a47.865 47.865 0 00-1.134-.175 2.31 2.31 0 01-1.64-1.055l-.822-1.316a2.192 2.192 0 00-1.736-1.039 48.774 48.774 0 00-5.232 0 2.192 2.192 0 00-1.736 1.039l-.821 1.316z"
+						/>
+						<path
+							stroke-linecap="round"
+							stroke-linejoin="round"
+							d="M16.5 12.75a4.5 4.5 0 11-9 0 4.5 4.5 0 019 0zM18.75 10.5h.008v.008h-.008V10.5z"
+						/>
+					</svg>
+					Receive via nostr
+				</div>
+				<QRCodeImage text={npub} scale={6} displayType="canvas" />
+			</div>
+		</div>
+	{/if}
+{/await}
