+++ src/comp/tokens/InboxTable.svelte
@@ -115,7 +115,7 @@
 					<p class="flex lg:hidden">Amt</p></th
 				>
 				<th>Date</th>
-				<th>From</th>
+				<th class="w-full">From</th>
 			</tr>
 		</thead>
 		<tbody class="max-h-1 overflow-y-scroll">
@@ -126,7 +126,7 @@
 			{/each}
 			<!-- svelte-ignore a11y-click-events-have-key-events -->
 			<tr class="">
-				<td colspan="2" class="cursor-pointer w-full hover:bg-base-200" on:click={loadMore}>
+				<td colspan="2" class="cursor-pointer w-5 hover:bg-base-200" on:click={loadMore}>
 					load more
 				</td>
 				<td
