+++ src/model/mint.ts
@@ -1,6 +1,8 @@
+import type { MintKeys } from "@cashu/cashu-ts/dist/lib/es5/model/types";
+
 interface Mint {
 	mintURL: string;
-	keys: object;
+	keys: MintKeys;
 	keysets: Array<string>;
 	isAdded: boolean;
 }
