+++ src/model/types/index.d.ts
@@ -1,0 +1,6 @@
+export {};
+declare global {
+  interface Window {
+    nostr: any;  // this will be your variable name
+  }
+}
